pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orbitlab"
version = "0.0.1"
dependencies = [
    "cryptography==46.0.3",
    "httpx==0.28.1",
    "pydantic==2.12.3",
    "pydantic-settings==2.11.0",
    "PyYAML==6.0.3",
    "reflex==0.8.18",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
]

[tool.ruff]
src = [ ".",]
line-length = 120
target-version = "py311"
extend-exclude = ["tests"]
output-format = "grouped"

[tool.ruff.lint]
select = [ "ALL" ]
ignore = [ "D203", "D212", "T201", "G004", "S507", "S202", "PTH207" ]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--capture=no --cov=pws --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
omit = [
    "*setup.py",
    "*tests/*",
    "*__global_coverage__/*",
]
